oct 20th update

+++ contactUpdateController.php
@@ -6,20 +6,20 @@
     //************************
     showErrors(0);  //1 - Turn on Error Display
 
-    // $method=$_SERVER['REQUEST_METHOD'];
+    $method=$_SERVER['REQUEST_METHOD'];
     // * Process HTTP GET Request
-    // if($method=='GET'){
+    if($method=='GET'){
         $contactDAO = new ContactDAO();
         $contactDAO->deleteContact(3);
 
-    // }
+    }
     
-    * Process HTTP POST Request
+    //* Process HTTP POST Request
     if($method=='POST'){
         $contact = new Contact();
         $contact -> contactID  =$_POST['contactID'];
         $contact -> username = $_POST['username'];
-        $contact -> email = $_POST['email']
+        $contact -> email = $_POST['email'];
 
         $contactDAO = new contactDAO();
